+++ backend/internal/handler/treatment_plan_request.go
@@ -5,7 +5,7 @@ type createTreatmentPlanRequest struct {
 	Memo             string  `json:"memo"`
 	Insurance        bool    `json:"insurance"`
 	UnitPrice        float64 `json:"unit_price"`
-	Quantity         int     `json:"quantity"`
+	Quantity         float64 `json:"quantity"`
 	DiscountRate     float64 `json:"discount_rate"`
 	DiscountAmount   float64 `json:"discount_amount"`
 	Subtotal         float64 `json:"subtotal"`
@@ -17,7 +17,7 @@ type updateTreatmentPlanRequest struct {
 	Memo             *string  `json:"memo"`
 	Insurance        *bool    `json:"insurance"`
 	UnitPrice        *float64 `json:"unit_price"`
-	Quantity         *int     `json:"quantity"`
+	Quantity         *float64 `json:"quantity"`
 	DiscountRate     *float64 `json:"discount_rate"`
 	DiscountAmount   *float64 `json:"discount_amount"`
 	Subtotal         *float64 `json:"subtotal"`
